+++ examples/adonis/app/middleware/container_bindings_middleware.ts
@@ -1,0 +1,9 @@
+import type { HttpContext } from "@adonisjs/core/http";
+import type { NextFn } from "@adonisjs/core/types/http";
+
+export default class ContainerBindingsMiddleware {
+  async handle(ctx: HttpContext, next: NextFn) {
+    ctx.containerResolver.bindValue(HttpContext, ctx);
+    return next();
+  }
+}