+++ examples/adonis/start/routes.ts
@@ -1,0 +1,9 @@
+import router from "@adonisjs/core/services/router";
+
+router.get("/", async () => {
+  return "Hello, Fedify!";
+});
+
+router.get("/users/:identifier", async ({ params }) => {
+  return `<p>Hello, ${params.identifier}!</p>`;
+});